+++ src/components/BoardStatusDialogue/index.ts
@@ -1,0 +1,1 @@
+export { BoardStatusDialogue } from "./BoardStatusDialogue";